% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreliminaryRtExtractor.R
\name{PreliminaryRtExtractor}
\alias{PreliminaryRtExtractor}
\title{Extract retention time from mzML data}
\usage{
PreliminaryRtExtractor(
  adductFile = NA,
  mzMLFile,
  MixedStandards,
  ppm = 10,
  haldWidth = 20,
  ignore_threshold = 0.1,
  span = 9
)
}
\description{
Extract retention time from mzML data
}
\author{
Fujian Zheng
}
